requirements.txt
# FastAPI + ASGI
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Async HTTP client (Instagram API calls)
httpx>=0.27.0

# JWT Authentication
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0

# Password hashing
bcrypt>=4.1.0
passlib[bcrypt]>=1.7.4

# Encryption (token storage)
cryptography>=42.0.0

# Azure Cosmos DB
azure-cosmos>=4.5.0

# Azure Key Vault (token encryption)
azure-keyvault-secrets>=4.7.0
azure-identity>=1.15.0

# Azure Service Bus (webhook message queue)
azure-servicebus>=7.12.0

# Redis
redis>=5.0.0

# Celery (task queue + scheduler)
celery[redis]>=5.3.0
kombu>=5.3.0

# Production server
gunicorn>=21.2.0
